fix circular dependancy warning
